requirements.txt
argcomplete
boto3
botocore
jmespath
requests
pyyaml
